Update CC_QuickSort.pde
It should be fixed now. The make the println into comments if you want. The error was in the exit condition: the right index was going lower than the left and you were always returning the left index even if you didn't swap.

Also on my computer, the rendering works fine the way you wrote it...